+++ lib/legion/extensions/microsoft_teams.rb
@@ -20,6 +20,7 @@ require 'legion/extensions/microsoft_teams/runners/cache_ingest'
 require 'legion/extensions/microsoft_teams/runners/people'
 require 'legion/extensions/microsoft_teams/runners/profile_ingest'
 require 'legion/extensions/microsoft_teams/runners/ownership'
+require 'legion/extensions/microsoft_teams/runners/api_ingest'
 
 # Helpers (bot)
 require 'legion/extensions/microsoft_teams/helpers/high_water_mark'