+++ lib/legion/extensions/microsoft_teams/actors/token_refresher.rb
@@ -24,7 +24,8 @@ module Legion
 
           def enabled?
             defined?(Legion::Extensions::MicrosoftTeams::Helpers::TokenCache)
-          rescue StandardError
+          rescue StandardError => e
+            log.debug("TokenRefresher#enabled?: #{e.message}")
             false
           end